backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
